Added my first subroutine ever for a new line. and it works! yay!!
The subroutine is the newLine subroutine. It is useful cuz we can print
everything on a new line without adding 7 lines of code just to move the
output one line down. btw it makes use of the stack. must learn how it
all works soon.
I also commented the code to print everything between the starting and
ending address