added the load stack pointer instruction
and an empty line for better appearance. Also, commented the code as one
should. make sure to comment as much as possible. good practice is to
comment each line almost right away even if it seems really obvious.
keep in mind, that it may be obvious to you but not to your teammates.
thanks.